HIVE-COTE refactor
