+++ src/main/java/experiments/SimulationExperiments.java
@@ -18,7 +18,7 @@ package experiments;
 //import com.sun.management.GarbageCollectorMXBean;
 import tsml.classifiers.dictionary_based.*;
 import tsml.classifiers.distance_based.DTWCV;
-import tsml.classifiers.hybrids.FlatCote;
+import tsml.classifiers.hybrids.legacy_cote.FlatCote;
 import tsml.classifiers.shapelet_based.LearnShapelets;
 import tsml.classifiers.shapelet_based.FastShapelets;
 import tsml.classifiers.interval_based.TSBF;
@@ -28,7 +28,7 @@ import tsml.classifiers.shapelet_based.ShapeletTransformClassifier;
 import tsml.classifiers.interval_based.LPS;
 import tsml.classifiers.distance_based.ElasticEnsemble;
 import tsml.classifiers.distance_based.DD_DTW;
-import tsml.classifiers.hybrids.HiveCote;
+import tsml.classifiers.hybrids.legacy_cote.HiveCote;
 import fileIO.InFile;
 import fileIO.OutFile;
 import java.io.File;
